Add integration test env variable
Add an environment variable to detect if the tests are integration
tests or not. This is required to store the id to be used on the
following tests. The order of the tests are:
- Get all entities (This test can be run in any order).
- Add entity (This creates a real new entity and returns the id that should be
	    used on the following tests).
- Get entity by id (Using the previous id).
- Update entity with id (Using the previous id).
- Delete entity with id (Using the previous id).

As you can see, the id is stored in order to have a real entity to be used on
the following tests.