Add tests-client script without webdriver
Add a gulp task for running the client tests without
auto-starting, auto-updating and auto-stoping the
webdriver. Having to wait so much meanwhile developing
tests isn't productive. Thus, there is an npm script for
starting the webdriver, and another for running the tests.

For CI, there is still a gulp task that takes care of
starting the webdriver, updating it, run the tests and stop
the webdriver.